Improve test coverage for DDIdAdapter
* Added a test exercising serialization of null DDId.
* Extract adapter to top-level class.
* Exclude DDIdAdapter from code coverage thresholds (even after the
  test, it won't pass the method coverage requirements).
