+++ dd-java-agent/agent-iast/src/test/groovy/com/datadog/iast/model/json/VulnerabilityEncodingTest.groovy
@@ -46,6 +46,37 @@ class VulnerabilityEncodingTest extends DDSpecification {
     }''')
   }
 
+  void 'one vulnerability with no spanId'() {
+    given:
+    final slurper = new JsonSlurper()
+    final value = new VulnerabilityBatch()
+    value.add(new Vulnerability(
+      VulnerabilityType.WEAK_HASH,
+      Location.forSpanAndStack(null, new StackTraceElement("foo", "foo", "foo", 1)),
+      new Evidence("MD5")
+      ))
+
+    when:
+    final result = VulnerabilityEncoding.toJson(value)
+
+    then:
+    slurper.parseText(result) == slurper.parseText('''{
+      "vulnerabilities": [
+        {
+          "type": "WEAK_HASH",
+          "evidence": {
+            "value": "MD5"
+          },
+          "hash":1042880134,
+          "location": {
+            "line": 1,
+            "path": "foo"
+          }
+        }
+      ]
+    }''')
+  }
+
   void 'one vulnerability with one source'() {
     given:
     final slurper = new JsonSlurper()
